Restructure main folders to clean up locations
