+++ burn_analyzer/config.py
